[CS-4106]: Update nav options on new navigator
